requirements.txt
arrow
speedtest-cli
